agentic_ai/requirements.txt
fastapi
uvicorn
langgraph>=0.2.0
langchain>=0.3.0
langchain-groq
langchain-community
langchain-text-splitters
pydantic>=2.0.0
sqlalchemy
pypdf
faiss-cpu
duckduckgo-search
python-dotenv
sentence-transformers